requirements.txt
Flask
pytesseract
PyMuPDF
pdf2image
Pillow
